Add support and test for complex variables

+++ diffrax/integrate.py
@@ -810,7 +810,9 @@ def diffeqsolve(
     save_index = 0
     made_jump = False if made_jump is None else made_jump
     ts = jnp.full(out_size, jnp.inf)
-    ys = jtu.tree_map(lambda y: jnp.full((out_size,) + jnp.shape(y), jnp.inf), y0)
+    ys = jtu.tree_map(
+        lambda y: jnp.full((out_size,) + jnp.shape(y), jnp.inf, dtype=y.dtype), y0
+    )
     result = jnp.array(RESULTS.successful)
     if saveat.dense:
         t0 = eqxi.error_if(t0, t0 == t1, "Cannot save dense output if t0 == t1")
